Introduce 'worker' object
This is really just an extension of the existing thread object. The idea
is to allocate threads and park them as 'idle' while waiting for work,
rather than spawning new threads and killing them constantly.

This object handles the state progression to ensure that the worker
thread can be started, stopped, and signaled.