Move client thread from proxy_conn to proxy itself
This may seem like an unnecessary complexity, but decoupling the thread
that's processing the connection to the client from the proxy_conn
connections makes it possible to decide which proxy_conn instance is
best based on authentication results.

This means there are now two pools maintained in the proxy: a pool of
proxy_conn instances, and a pool of proxy_worker instances.